feat: fullpic 라우팅, 폴더구조 정리
